backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
python-dotenv>=1.0.0
djangorestframework-simplejwt>=5.2.2
Pillow>=9.0.0
boto3>=1.26.0
django-storages>=1.13.2

# Voice Agent dependencies
livekit-agents>=0.8.0
livekit-plugins-openai>=0.6.0
livekit-plugins-deepgram>=0.6.0
livekit-plugins-silero>=0.6.0
livekit-api>=0.6.0